fix: types, id assignment
